requirements.txt
jsonschema
matplotlib
numpy
parse
scipy
sklearn
typing
sacred
sqlalchemy
pyyaml
progressbar2
spacy==1.9
diskcache
pytest
pytest-runner
pytest-xdist
pytest-pep8
pytest-xdist
pytest-cov
codecov
